requirements.txt
binance==0.3
websocket_client==0.57.0
